seperate files and add debug msg (WIP)
