+++ index.js
@@ -1,140 +1,48 @@
 const { readFileSync, writeFileSync } = require('fs')
-const cors = require('cors')
 const path = require('path').resolve()
 const https = require('https')
 const express = require('express')
-const session = require('express-session')
-const MemoryStore = require('memorystore')(session)
-const bodyParser = require('body-parser')
-const { Client } = require('discord.js')
-const DiscordOAuth2 = require('discord-oauth2')
+const debug = require('debug')('neptunebot')
 
-const authCheck = require('./auth')
 // const pointModule = require('./modules/point')
-const commandModule = require('./modules/command')
-const webRouter = require('./modules/router')
 const dataMgr = require('./modules/dataManager')
-
 const settings = require(path + '/settings.json')
 const authData = require(path + '/auth/authData.json')
+
+debug('starting NeptuneBot')
+
+debug('loading discord bot')
+const bot = require('./bot')
+
+debug('loading web server')
+const web = require('./web')
 
 if (!settings.features) settings.features = {}
 if (settings.features.serverStatus == null) settings.features.serverStatus = true
 if (settings.features.verifiedCheck == null) settings.features.verifiedCheck = true
-
-const authUrl = 'https://discordapp.com/api/oauth2/authorize?client_id=' +
-  settings.auth.clientId + '&redirect_uri=' +
-  encodeURI(settings.auth.redirectUri) +
-  '&response_type=code&scope=' + settings.auth.scope
-
-const app = express()
-const bot = new Client()
 
 bot.settings = settings
 
 let ssl
 if (!settings.development) ssl = { cert: readFileSync(path + '/auth/teaddy-cert.pem'), key: readFileSync(path + '/auth/teaddy-key.pem') }
 
-const discordOAuth = new DiscordOAuth2()
-
 dataMgr.startAutosave()
 
-// Web
-app.set('view engine', 'ejs')
-app.set('views', './page')
-
-app.use(session({
-  name: 'teaddy-session',
-  resave: false,
-  saveUninitialized: false,
-  secret: 'keyboard cat', // TODO load from settings.json
-  store: new MemoryStore({
-    checkPeriod: 1000 * 3600 // 1 hour
-  })
-}))
-
-app.use(bodyParser.urlencoded({ extended: false }))
-
 // Include data to req object
-app.use((req, _res, next) => {
+web.use((req, _res, next) => {
   req.dataMgr = dataMgr
   next()
 })
 
-app.use(cors())
-app.use('/src', express.static(path + '/src'))
-
-// Main Router (WIP)
-app.use(webRouter)
-
-// Old Login Page
-app.get('/oldlogin', async (req, res) => {
-  let key = req.query.key ? req.query.key.split(';') : []
-  let discordData = {}
-  try {
-    discordData = await discordOAuth.getUser(key[0])
-  } catch (err) {
-    key = []
-  }
-  res.render('oldlogin', { key, authUrl, authData, discordData })
-})
-
-app.get('/solve', (_req, res) => res.send({ items: ['discord', 'google'] }))
-app.get('/solve/:item', (req, res) => {
-  const { item } = req.params
-  let code = req.query.code || ''
-
-  code = code.split(';')
-
-  if (code[0].length <= 0) res.redirect('/oldlogin')
-  else {
-    switch (item) {
-      case 'discord':
-        authCheck.discord(settings.auth, code, discordOAuth).then((returnData) => {
-          authData[returnData.token] = { discord: returnData.userData, verified: false }
-          res.redirect('/oldlogin?key=' + returnData.token)
-        }).catch((err) => {
-          console.error(err)
-          res.sendStatus(500)
-        })
-        break
-
-      case 'google':
-        if (code.length !== 2) res.redirect('/oldlogin')
-        if (!Object.keys(authData).includes(code[0])) res.redirect('/login')
-        else {
-          authCheck.google(code[1]).then((data) => {
-            if (!data) res.sendStatus(401)
-            else {
-              authData[code[0]].google = data.body
-              authData[code[0]].verified = false
-              res.redirect('/oldlogin?key=' + code[0] + ';' + code[1])
-            }
-          }).catch(() => { res.sendStatus(401) })
-        }
-        break
-      case 'submit':
-        if (code.length !== 2) res.sendStatus(400)
-        else if (!(authData[code[0]] && authData[code[0]].discord && authData[code[0]].google)) res.sendStatus(401)
-        else {
-          authData[code[0]].verified = true
-          bot.channels.get(settings.channelId)
-            .send('<@' + authData[code[0]].discord.id + '>님의 인증이 완료되었습니다.')
-          res.redirect(settings.inviteUrl)
-        }
-        break
-    }
-  }
-})
-
 let server
-if (settings.development) server = app
-else server = https.createServer(ssl, app)
+if (settings.development) server = web
+else server = https.createServer(ssl, web)
 server.listen(settings.port, () => {
   console.log('Neptune Bot is not running on http://localhost:' + settings.port)
 })
 
 bot.login(settings.token)
+/*
   .then(() => {
     if (settings.features.verifiedCheck) {
       console.log('Verification Check Enabled.')
@@ -160,21 +68,7 @@ bot.login(settings.token)
       }, 1000)
     }
   })
+  */
 
 // TODO Delete
 setInterval(() => { writeFileSync(path + '/auth/authData.json', JSON.stringify(authData)) }, 1000)
-
-// Register Events
-bot.once('ready', () => {
-  commandModule.init(bot)
-})
-
-bot.on('message', (msg) => {
-  if (msg.author.bot || !msg.content) return
-  if (msg.content.startsWith(settings.prefix || '!')) {
-    // Commands
-    commandModule.run(msg, bot)
-  } else {
-    // Chatting Point
-  }
-})
