+++ modules/discordAuth.js
@@ -2,10 +2,7 @@ const DiscordOAuth2 = require('discord-oauth2')
 const OAuth = new DiscordOAuth2()
 
 function loginURL (settings) {
-  return 'https://discord.com/api/oauth2/authorize?client_id=' +
-  settings.clientId + '&redirect_uri=' +
-  encodeURI(settings.redirectUri) +
-  '&response_type=code&scope=identify'
+  return `https://discord.com/api/oauth2/authorize?client_id=${settings.clientId}&redirect_uri=${encodeURI(settings.redirectUri)}&response_type=code&scope=identify`
 }
 
 async function authorize (authData, code) {
