+++ package.json
@@ -4,7 +4,9 @@
   "description": "Member Management Bot for Teaddy Studio Discord Server",
   "main": "index.js",
   "scripts": {
-    "test": "node index"
+    "test": "node index",
+    "debug": "DEBUG=neptunebot* node .",
+    "debug:all": "DEBUG=* node ."
   },
   "repository": {
     "type": "git",
@@ -33,16 +35,17 @@
     "eslint-plugin-standard": "^4.0.1"
   },
   "dependencies": {
-    "body-parser": "^1.19.0",
     "cors": "^2.8.5",
+    "debug": "^4.1.1",
     "discord-oauth2": "^2.2.0",
-    "discord.js": "^11.5.1",
-    "ejs": "^3.0.1",
+    "discord.js": "^12.3.1",
+    "ejs": "^3.1.3",
     "express": "^4.17.1",
     "express-session": "^1.17.1",
-    "flake-idgen": "^1.4.0",
-    "js-sha256": "^0.9.0",
     "memorystore": "^1.6.2",
     "superagent": "^5.2.1"
+  },
+  "nodemonConfig": {
+    "ignore": ["auth/*", "data/*"]
   }
 }
