+++ page/_head.ejs
@@ -20,7 +20,7 @@
       <li id="topbar_account">
         <span id="topbar_account_item" class="topbar-item nav-dropdown-keep-show">계정</span>
         <div id="topbar_account_content" class="nav-dropdown-content nav-dropdown-keep-show">
-          <% if(!id) { %>
+          <% if(!user) { %>
           <a href="./login" id="topbar_account_login" class="card">
             <span>다양한 기능을 이용하시려면 <b><u>로그인</u></b>하세요.</span>
           </a>
