+++ page/index.ejs
@@ -1,2 +1,2 @@
-<%- include('./_head.ejs', { id, page: 'index' }) %>
+<%- include('./_head.ejs', { user, page: 'index' }) %>
 <%- include('./_tail.ejs') %>
